fix/move block.rs to lib.rs
